test_requirements.txt
-e .
dj-database-url==0.3.0
djangorestframework==2.4.2
factory-boy==2.4.1
psycopg2==2.5.3
py==1.4.23
pytest==2.6.1
pytest-cov==1.8.0
pytest-django==2.6.2
pytest-pythonpath==0.3.0
